rag model/requirements.txt
# Core dependencies
sqlalchemy>=2.0.44
psycopg2-binary>=2.9.11

# PDF processing
pypdf2>=3.0.1
pdfplumber>=0.11.8
pdf2image>=1.17.0
pytesseract>=0.3.13
pillow>=11.2.1

# OCR dependencies (Tesseract)
# Note: Tesseract OCR engine needs to be installed separately
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Linux: sudo apt-get install tesseract-ocr
# Mac: brew install tesseract

#Poppler utils for pdf2image
# Note: Poppler needs to be installed separately
# Windows: https://github.com/oschwartz10612/poppler-windows/releases/tag/v25.11.0-0

# NLP and text processing
nltk>=3.9.2
numpy>=2.2.3

# Modern Python alternatives (optional)
pydantic>=2.12.3
pydantic-settings>=2.12.0

# Utility libraries
python-dotenv>=1.2.1

# Core Dependencies
python>=3.13.2

# PDF Processing
pdfminer.six>=20221107

# Text Processing & NLP
regex>=2025.10.23

# Vector Embeddings & Machine Learning
sentence-transformers>=5.1.2
torch>=2.9.0
transformers>=4.57.1

# Vector Database
pinecone-client>=6.0.0

# LLM & AI Services
google-generativeai>=0.8.5
#openai==1.3.0

# LangChain Ecosystem
langchain>=1.0.7
langchain-community>=0.4.1
langchain-google-genai>=3.0.3
langchain-core>=1.0.5
langchain-text-splitters>=1.0.0

# Web & API
requests>=2.32.5
urllib3>=2.3.0

# Data Processing & Utilities
pandas>=2.3.0


# Additional Utilities
tqdm>=4.67.1
colorama>=0.4.6
pyyaml>=6.0.2

# For monitoring
langsmith>=0.4.43